Determine dtype for summary files
The numpy structured array is a perfect fit for the data in a summary
file, which is really just a set of column vectors.

The problem is complicated slightly by the files somtimes containing
void columns, which are not to be exposed, at least not through the main
interface.

The implementation is an early prototype, to get tests up and running.
It is fairly slow and inelegant, but is quite straight forward.

The Summary.pos attribute is a bit awkward, and likely to change in the
future.