Add Summary.readall
The readall function is first and foremost convenient for the quite
common usecase of reading all data from a summary file into memory. It
works on files and not on cases, and assumes local filesystem only for
now, through anything iostream can open.

It adds a chunk of code to the ecl3/io.hpp file which is already now
destinated to either die or undergo substantial changes - as such,
comments in code emphasise that this is *not* stable and cannot be
relied on as such, as opposed to parts of the C interface.

What it does, accomplish, however, is showcase the use of structured
numpy arrays and fully qualified columns.