Rename ecl3.summary.smspec -> ecl3.summary.summary
As the design has developed, it becomes clearer that the summary class
actually represents the concept of a summary more than just the
specification file. Granted, all the attributes of the object are lifted
from the summary, but that is simply a consequence of how data is
organised anyway.

The publically-exported smspec name was probably always wrong, and
summary seems more natural.